+++ protoc_plugin/CHANGELOG.md
@@ -4,6 +4,7 @@
   corresponding .proto file.
 
 ## 16.0.3
+
 * Sync Kythe metadata updates from internal repo:
 * Remove the extra 1 (field name) from generated field paths,
   so they refer to the whole field rather than the name.